move scanner to separate package
